address comments 2 -- move getting server uri for segments to executor instead
