+++ pinot-plugins/pinot-minion-tasks/pinot-minion-builtin-tasks/src/main/java/org/apache/pinot/plugin/minion/tasks/BaseTaskExecutor.java
@@ -20,10 +20,6 @@ package org.apache.pinot.plugin.minion.tasks;
 
 import com.google.common.base.Preconditions;
 import java.io.File;
-import java.net.URI;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import org.apache.commons.io.FileUtils;
 import org.apache.pinot.common.metadata.ZKMetadataProvider;
 import org.apache.pinot.common.metadata.segment.SegmentZKMetadata;
@@ -32,6 +28,7 @@ import org.apache.pinot.common.metrics.MinionMeter;
 import org.apache.pinot.common.metrics.MinionMetrics;
 import org.apache.pinot.common.utils.fetcher.SegmentFetcherFactory;
 import org.apache.pinot.core.minion.PinotTaskConfig;
+import org.apache.pinot.core.util.PeerServerSegmentFinder;
 import org.apache.pinot.minion.MinionContext;
 import org.apache.pinot.minion.executor.PinotTaskExecutor;
 import org.apache.pinot.spi.config.table.TableConfig;
@@ -111,7 +108,7 @@ public abstract class BaseTaskExecutor implements PinotTaskExecutor {
   }
 
   protected void downloadSegmentToLocal(String tableNameWithType, String segmentName, String deepstoreURL,
-      Map<String, List<String>> segmentServiceUrisList, File tarredSegmentFile, String crypterName)
+      String taskType, File tarredSegmentFile, String crypterName)
       throws Exception {
     LOGGER.info("Downloading segment from {} to {}", deepstoreURL, tarredSegmentFile.getAbsolutePath());
     try {
@@ -119,12 +116,15 @@ public abstract class BaseTaskExecutor implements PinotTaskExecutor {
       SegmentFetcherFactory.fetchAndDecryptSegmentToLocal(deepstoreURL, tarredSegmentFile, crypterName);
     } catch (Exception e) {
       LOGGER.error("Segment download failed from deepstore for {}, crypter:{}", deepstoreURL, crypterName, e);
-      if (!segmentServiceUrisList.isEmpty()) {
+      TableConfig tableConfig = getTableConfig(tableNameWithType);
+      String peerDownloadScheme = tableConfig.getValidationConfig().getPeerSegmentDownloadScheme();
+      if (MinionTaskUtils.extractMinionAllowDownloadFromServer(tableConfig, taskType) && peerDownloadScheme != null) {
         LOGGER.info("Trying to download from servers for segment {} post deepstore download failed", segmentName);
         SegmentFetcherFactory.getSegmentFetcher(
                 getTableConfig(tableNameWithType).getValidationConfig().getPeerSegmentDownloadScheme())
             .fetchSegmentToLocal(segmentName, () ->
-                    segmentServiceUrisList.get(segmentName).stream().map(URI::create).collect(Collectors.toList()),
+                PeerServerSegmentFinder.getPeerServerURIs(MINION_CONTEXT.getHelixManager(),
+                    tableNameWithType, segmentName, peerDownloadScheme),
                 tarredSegmentFile);
       } else {
         throw e;
