+++ pinot-plugins/pinot-minion-tasks/pinot-minion-builtin-tasks/src/main/java/org/apache/pinot/plugin/minion/tasks/MinionTaskUtils.java
@@ -18,9 +18,6 @@
  */
 package org.apache.pinot.plugin.minion.tasks;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -33,8 +30,9 @@ import org.apache.pinot.common.restlet.resources.ValidDocIdsBitmapResponse;
 import org.apache.pinot.common.utils.config.InstanceUtils;
 import org.apache.pinot.controller.helix.core.minion.ClusterInfoAccessor;
 import org.apache.pinot.controller.util.ServerSegmentMetadataReader;
-import org.apache.pinot.core.common.MinionConstants;
 import org.apache.pinot.minion.MinionContext;
+import org.apache.pinot.spi.config.table.TableConfig;
+import org.apache.pinot.spi.config.table.TableTaskConfig;
 import org.apache.pinot.spi.env.PinotConfiguration;
 import org.apache.pinot.spi.filesystem.LocalPinotFS;
 import org.apache.pinot.spi.filesystem.PinotFS;
@@ -52,7 +50,6 @@ public class MinionTaskUtils {
   private static final Logger LOGGER = LoggerFactory.getLogger(MinionTaskUtils.class);
 
   private static final String DEFAULT_DIR_PATH_TERMINATOR = "/";
-  private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
   private MinionTaskUtils() {
   }
@@ -191,15 +188,18 @@ public class MinionTaskUtils {
     return servers;
   }
 
-  public static Map<String, List<String>> getSegmentServerUrisList(Map<String, String> configs)
-      throws JsonProcessingException {
-    String segmentServerURIs = configs.getOrDefault(MinionConstants.SEGMENT_SERVER_URIS_LIST_KEY, "");
-    Map<String, List<String>> segmentServerUrisList = new HashMap<>();
-    if (!segmentServerURIs.equals("")) {
-      segmentServerUrisList = OBJECT_MAPPER.readValue(configs.get(MinionConstants.SEGMENT_SERVER_URIS_LIST_KEY),
-          new TypeReference<>() {
-          });
+  /**
+   * Extract allowDownloadFromServer config from table task config
+   */
+  public static boolean extractMinionAllowDownloadFromServer(TableConfig tableConfig, String taskType) {
+    TableTaskConfig tableTaskConfig = tableConfig.getTaskConfig();
+    if (tableTaskConfig != null) {
+      Map<String, String> configs = tableTaskConfig.getConfigsForTaskType(taskType);
+      if (configs != null && !configs.isEmpty()) {
+        return Boolean.parseBoolean(configs.getOrDefault(TableTaskConfig.MINION_ALLOW_DOWNLOAD_FROM_SERVER,
+            String.valueOf(TableTaskConfig.DEFAULT_MINION_ALLOW_DOWNLOAD_FROM_SERVER)));
+      }
     }
-    return segmentServerUrisList;
+    return TableTaskConfig.DEFAULT_MINION_ALLOW_DOWNLOAD_FROM_SERVER;
   }
 }
